Rename namespce of MockFastpathConnection test classes
